Began implementing section table view controller and add section view controller.

+++ rED/Notebook.h
@@ -34,7 +34,7 @@
 // Object Storage
 @property (nonatomic, strong) NSMutableArray *array_sections;
 @property (nonatomic, strong) NSMutableArray *array_highlights;
-@property int indexOfLastLoadedSection;
+@property NSUInteger indexOfLastLoadedSection;
 
 // Initializers
 - (instancetype)initWithCoder:(NSCoder *)aDecoder;
